+++ src/pages/Cms/components/layout/CenterContent/index.tsx
@@ -9,7 +9,7 @@ const Component: React.FC = (): JSX.Element => {
   const dropRef = useRef(null)
   const handleDrop = (event: any) => {
     const data = utils.JSONPase(event.dataTransfer.getData('text/plain'))
-    setSourceData([{ type: data.type }, ...sourceData])
+    setSourceData([...sourceData, { type: data.type }])
   }
   const handleDragOver = (event: any) => {
     event.preventDefault()